+++ src/constants/settings.ts
@@ -40,6 +40,7 @@ export const SETTINGS_CONTENT_STATIC_FOLDERS = "content.publicFolder";
 export const SETTINGS_CONTENT_FRONTMATTER_HIGHLIGHT = "content.fmHighlight";
 
 export const SETTINGS_DASHBOARD_OPENONSTART = "dashboard.openOnStart";
+export const SETTINGS_DASHBOARD_MEDIA_SNIPPET = "dashboard.mediaSnippet";
 
 /**
  * @deprecated